# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/catalog/medicines": {
      "get": {
        "operationId": "CatalogController_search",
        "parameters": [
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Search text (min 2 chars); omit to list",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "zoneId",
            "required": false,
            "in": "query",
            "description": "Restrict to in-stock items in this delivery zone",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "List/search medicines (fuzzy: tolerates misspellings like \"dollo\"). Pass zoneId to only see what is in stock nearby, with the cheapest in-zone price.",
        "tags": [
          "catalog"
        ]
      }
    },
    "/v1/catalog/medicines/{id}": {
      "get": {
        "operationId": "CatalogController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "zoneId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Medicine detail; with zoneId includes which in-zone shops stock it",
        "tags": [
          "catalog"
        ]
      }
    },
...
    "/v1/orders": {
      "post": {
        "operationId": "OrdersController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrderDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Place an order. Server re-prices everything; Razorpay orders return checkout params and stay PENDING until the payment webhook confirms.",
        "tags": [
          "orders"
        ]
      },
      "get": {
        "operationId": "OrdersController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "My orders (latest first)",
        "tags": [
          "orders"
        ]
      }
    },
    "/v1/orders/{id}": {
      "get": {
        "operationId": "OrdersController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "My order detail — includes the delivery code to hand to the rider",
        "tags": [
          "orders"
        ]
      }
    },
    "/v1/orders/{id}/cancel": {
      "post": {
        "operationId": "OrdersController_cancel",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CancelDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Cancel my order (allowed until the shop packs it); prepaid amounts auto-refund",
        "tags": [
          "orders"
        ]
      }
    },
    "/v1/shop/orders": {
      "get": {
        "operationId": "ShopOrdersController_list",
        "parameters": [
          {
            "name": "state",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "PLACED",
                "RX_REVIEW",
                "ACCEPTED",
                "PACKED",
                "RIDER_ASSIGNED",
                "PICKED_UP",
                "OUT_FOR_DELIVERY",
                "DELIVERED",
                "RX_REJECTED",
                "CANCELLED",
                "UNDELIVERED"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Orders for my shop (unpaid online orders are hidden until payment lands)",
        "tags": [
          "shop-portal"
        ]
      }
    },
    "/v1/shop/orders/{id}": {
      "get": {
        "operationId": "ShopOrdersController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Order detail for my shop",
        "tags": [
          "shop-portal"
        ]
      }
    },
    "/v1/shop/orders/{id}/accept": {
      "post": {
        "operationId": "ShopOrdersController_accept",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ShopAcceptDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Confirm availability item-by-item. Dropped items auto-refund; zero available items cancels the order.",
        "tags": [
          "shop-portal"
        ]
      }
    },
    "/v1/shop/orders/{id}/pack": {
      "post": {
        "operationId": "ShopOrdersController_pack",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Mark the order packed and ready for rider pickup",
        "tags": [
          "shop-portal"
        ]
      }
    },
    "/v1/admin/orders": {
      "get": {
        "operationId": "AdminOrdersController_list",
        "parameters": [
          {
            "name": "state",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "PLACED",
                "RX_REVIEW",
                "ACCEPTED",
                "PACKED",
                "RIDER_ASSIGNED",
                "PICKED_UP",
                "OUT_FOR_DELIVERY",
                "DELIVERED",
                "RX_REJECTED",
                "CANCELLED",
                "UNDELIVERED"
              ],
              "type": "string"
            }
          },
          {
            "name": "shopId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Live orders board data",
        "tags": [
          "admin"
        ]
      }
    },
    "/v1/admin/orders/{id}": {
      "get": {
        "operationId": "AdminOrdersController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Full order detail incl. payments, refunds and audit trail",
        "tags": [
          "admin"
        ]
      }
    },
    "/v1/admin/orders/{id}/assign-rider": {
      "post": {
        "operationId": "AdminOrdersController_assignRider",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignRiderDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Assign (or reassign) a rider to a packed order",
        "tags": [
          "admin"
        ]
      }
    },
    "/v1/admin/orders/{id}/transition": {
      "post": {
        "operationId": "AdminOrdersController_transition",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminTransitionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Ops manual override (phone-confirmed acceptance, forced cancel, etc.). Obeys the state machine; refunds/COD side effects apply automatically.",
        "tags": [
          "admin"
        ]
      }
    },
    "/v1/rider/tasks": {
      "get": {
        "operationId": "RiderOrdersController_tasks",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "My active delivery tasks (assigned → out for delivery)",
        "tags": [
          "rider"
        ]
      }
    },
    "/v1/rider/orders/{id}/pickup": {
      "post": {
        "operationId": "RiderOrdersController_pickup",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Picked the parcel up from the shop",
        "tags": [
          "rider"
        ]
      }
    },
...
  },
  "info": {
    "title": "MediLocal API",
    "description": "Hyperlocal medicine delivery platform API. The OpenAPI spec at /docs-json also feeds the Dart client generator for the Flutter apps.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "OrderItemInputDto": {
        "type": "object",
        "properties": {
          "medicineId": {
            "type": "string"
          },
          "qty": {
            "type": "number",
            "minimum": 1,
            "maximum": 50
          }
        },
        "required": [
          "medicineId",
          "qty"
        ]
      },
      "CreateOrderDto": {
        "type": "object",
        "properties": {
          "addressId": {
            "type": "string",
            "description": "One of my saved addresses (its zone decides fees & the serving shop)"
          },
          "shopId": {
            "type": "string",
            "description": "Pin a shop; omit to let the platform pick (same rule as cart quote)"
          },
          "paymentMethod": {
            "type": "string",
            "enum": [
              "RAZORPAY",
              "COD"
            ]
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderItemInputDto"
            }
          },
          "rxFileKeys": {
            "description": "S3 file keys from POST /v1/prescriptions/upload-url — required when any item needs a prescription",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "addressId",
          "paymentMethod",
          "items"
        ]
      },
      "CancelDto": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string"
          }
        }
      },
      "ItemDecisionDto": {
        "type": "object",
        "properties": {
          "orderItemId": {
            "type": "string"
          },
          "accepted": {
            "type": "boolean",
            "description": "false = shop does not have it → item is dropped and auto-refunded"
          }
        },
        "required": [
          "orderItemId",
          "accepted"
        ]
      },
      "ShopAcceptDto": {
        "type": "object",
        "properties": {
          "items": {
            "description": "A decision for every item on the order",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ItemDecisionDto"
            }
          }
        },
        "required": [
          "items"
        ]
      },
      "AssignRiderDto": {
        "type": "object",
        "properties": {
          "riderId": {
            "type": "string"
          }
        },
        "required": [
          "riderId"
        ]
      },
      "AdminTransitionDto": {
        "type": "object",
        "properties": {
          "toState": {
            "type": "string",
            "enum": [
              "PLACED",
              "RX_REVIEW",
              "ACCEPTED",
              "PACKED",
              "RIDER_ASSIGNED",
              "PICKED_UP",
              "OUT_FOR_DELIVERY",
              "DELIVERED",
              "RX_REJECTED",
              "CANCELLED",
              "UNDELIVERED"
            ]
          },
          "note": {
            "type": "string",
            "description": "Why ops is overriding — goes into the order audit trail"
          }
        },
        "required": [
          "toState",
          "note"
        ]
      },
...
    }
  }
}